Move cube models to gammapy.modeling
