+++ gammapy/cube/fit.py
@@ -5,10 +5,15 @@ import astropy.units as u
 from astropy.io import fits
 from astropy.nddata.utils import NoOverlapError
 from astropy.utils import lazyproperty
-from gammapy.cube.models import BackgroundModel, BackgroundModels, SkyModel, SkyModels
 from gammapy.cube.psf_kernel import PSFKernel
 from gammapy.irf import EnergyDispersion
 from gammapy.maps import Map
+from gammapy.modeling.models import (
+    BackgroundModel,
+    BackgroundModels,
+    SkyModel,
+    SkyModels,
+)
 from gammapy.stats import cash, cash_sum_cython, cstat, cstat_sum_cython
 from gammapy.utils.fitting import Dataset, Parameters
 from gammapy.utils.random import get_random_state
