+++ gammapy/cube/simulate.py
@@ -7,8 +7,8 @@ from gammapy.cube import (
     make_map_background_irf,
     make_map_exposure_true_energy,
 )
-from gammapy.cube.models import BackgroundModel
 from gammapy.maps import WcsNDMap
+from gammapy.modeling.models import BackgroundModel
 from gammapy.utils.random import get_random_state
 
 __all__ = ["simulate_dataset"]
