+++ gammapy/cube/tests/test_simulate.py
@@ -4,10 +4,9 @@ from numpy.testing import assert_allclose
 import astropy.units as u
 from astropy.coordinates import SkyCoord
 from gammapy.cube import MapDataset, simulate_dataset
-from gammapy.cube.models import SkyModel, SkyModels
 from gammapy.irf import load_cta_irfs
 from gammapy.maps import MapAxis, WcsGeom
-from gammapy.modeling.models import PowerLaw, SkyGaussian
+from gammapy.modeling.models import PowerLaw, SkyGaussian, SkyModel, SkyModels
 from gammapy.utils.testing import requires_data
