+++ gammapy/modeling/models/__init__.py
@@ -1,5 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Built-in models in Gammapy."""
+from .cube import *
 from .image import *
 from .spectrum import *
 from .time import *
