+++ gammapy/modeling/models/cube/__init__.py
@@ -1,0 +1,3 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""Cube models."""
+from .core import *
