+++ gammapy/modeling/models/cube/core.py
@@ -4,7 +4,6 @@ from pathlib import Path
 import numpy as np
 import astropy.units as u
 from gammapy.maps import Map
-from gammapy.modeling.models import SkySpatialModel, SpectralModel
 from gammapy.utils.fitting import Model, Parameter, Parameters
 from gammapy.utils.scripts import make_path, write_yaml
 
@@ -205,6 +204,8 @@ class SkyModel(SkyModelBase):
     __slots__ = ["name", "_spatial_model", "_spectral_model"]
 
     def __init__(self, spatial_model, spectral_model, name="source"):
+        from gammapy.modeling.models import SkySpatialModel, SpectralModel
+
         self.name = name
 
         if not isinstance(spatial_model, SkySpatialModel):
