+++ gammapy/spectrum/tests/test_flux_point_estimator.py
@@ -5,10 +5,14 @@ from numpy.testing import assert_allclose
 from astropy import units as u
 from astropy.coordinates import SkyCoord
 from gammapy.cube import simulate_dataset
-from gammapy.cube.models import SkyModel
 from gammapy.irf import EffectiveAreaTable, load_cta_irfs
 from gammapy.maps import MapAxis, WcsGeom
-from gammapy.modeling.models import ExponentialCutoffPowerLaw, PowerLaw, SkyGaussian
+from gammapy.modeling.models import (
+    ExponentialCutoffPowerLaw,
+    PowerLaw,
+    SkyGaussian,
+    SkyModel,
+)
 from gammapy.spectrum import (
     FluxPointsEstimator,
     SpectrumDatasetOnOff,
