+++ gammapy/utils/random/tests/test_inverse_cdf.py
@@ -7,9 +7,13 @@ from astropy.coordinates import SkyCoord
 from astropy.table import Table
 from astropy.time import Time
 from gammapy.cube import MapEvaluator
-from gammapy.cube.models import SkyModel
 from gammapy.maps import Map, MapAxis
-from gammapy.modeling.models import LightCurveTableModel, PowerLaw, SkyGaussian
+from gammapy.modeling.models import (
+    LightCurveTableModel,
+    PowerLaw,
+    SkyGaussian,
+    SkyModel,
+)
 from gammapy.utils.random import InverseCDFSampler, MapEventSampler
