+++ gammapy/utils/serialization/io.py
@@ -1,14 +1,15 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Utilities to serialize models."""
 from gammapy.cube.fit import MapDataset, MapEvaluator
-from gammapy.cube.models import (
+from gammapy.modeling.models import (
+    SPATIAL_MODELS,
+    SPECTRAL_MODELS,
     BackgroundModel,
     BackgroundModels,
     SkyDiffuseCube,
     SkyModel,
     SkyModels,
 )
-from gammapy.modeling.models import SPATIAL_MODELS, SPECTRAL_MODELS
 from gammapy.utils.fitting import Parameters
 
 __all__ = ["models_to_dict", "dict_to_models", "dict_to_datasets", "datasets_to_dict"]
