+++ gammapy/utils/serialization/tests/test_io.py
@@ -2,7 +2,7 @@
 import numpy as np
 from numpy.testing import assert_allclose
 from astropy.utils.data import get_pkg_data_filename
-from gammapy.cube.models import BackgroundModels, SkyModels
+from gammapy.modeling.models import BackgroundModels, SkyModels
 from gammapy.modeling.models import image as spatial
 from gammapy.modeling.models import spectrum as spectral
 from gammapy.utils.fitting import Datasets
