+++ gammapy/utils/serialization/tests/test_xml.py
@@ -2,7 +2,7 @@
 import pytest
 import numpy as np
 from numpy.testing import assert_allclose
-from gammapy.cube.models import SkyModels
+from gammapy.modeling.models import SkyModels
 from gammapy.modeling.models import image as spatial
 from gammapy.modeling.models import spectrum as spectral
 from gammapy.utils.serialization import UnknownModelError, xml_to_sky_models
