+++ gammapy/utils/serialization/xml.py
@@ -9,9 +9,9 @@ For XML model format definitions, see here:
 import logging
 import numpy as np
 import astropy.units as u
-from gammapy.cube.models import SkyModel, SkyModels
 from gammapy.extern import xmltodict
 from gammapy.maps import Map
+from gammapy.modeling.models import SkyModel, SkyModels
 from gammapy.modeling.models import image as spatial
 from gammapy.modeling.models import spectrum as spectral
 from gammapy.utils.fitting import Parameter, Parameters
